Split launch-at-login handlers out of settings_presenter.rs
settings_presenter.rs was at 1115 lines, over the 1000-line structural CI
gate. Extract the three launch-at-login functions (user-event dispatcher,
set-state action, startup snapshot) into settings_presenter_launch_at_login.rs
following the existing sibling-module pattern used for backup/mcp/tool_approval
handlers. Drops settings_presenter.rs to 965 lines. No behavior change.
