+++ src/presentation/mod.rs
@@ -51,6 +51,7 @@ pub mod model_selector_presenter;
 pub mod profile_editor_presenter;
 pub mod settings_presenter;
 mod settings_presenter_backup;
+mod settings_presenter_launch_at_login;
 mod settings_presenter_mcp;
 mod settings_presenter_tool_approval;
 pub mod view_command;
